\name{lmeNBBayes-internal}
\title{Internal lmeNBBayes functions}
\alias{print.LinearMixedEffectNBBayes}
\alias{print.IndexBatch}

\description{Internal lmeNBBayes functions for printing}
\usage{
\method{print}{LinearMixedEffectNBBayes}(x,...)
\method{print}{IndexBatch}(x,...)
}
\author{Yumi Kondo <y.kondo@stat.ubc.ca>}
\keyword{internal}




